massive refactoring and moving of files
