+++ client/src/components/articles/Embeddings.js
