+++ client/src/components/authModal/AuthenticationModal.js
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import Validator from "jsonschema";
 import { Button, Header, Modal } from "semantic-ui-react";
 
-import history from "../utils/History";
+import history from "../../utils/History";
 import { AuthSchema } from "../common/Schemas";
 
 class AuthenticationModal extends React.Component {
