+++ client/src/components/common/CommonStyles.js
