+++ client/src/components/common/css/content-styles.css
