+++ client/src/index.js
@@ -5,7 +5,7 @@ import * as serviceWorker from "./serviceWorker";
 import axios from "axios";
 import "gun/sea";
 import "semantic-ui-css/semantic.min.css";
-import "./common/css/content-styles.css";
+import "./components/common/css/content-styles.css";
 import "@pathofdev/react-tag-input/build/index.css";
 require("dotenv").config();
